requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.28.0
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==10.0.0
celery==5.5.0
redis==5.2.1
django-redis==6.0.0
django-extensions==3.2.3

# Optional but recommended
gunicorn==22.0.0
whitenoise==6.8.2
django-storages==1.14.4
boto3==1.35.85 